requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
gunicorn==21.2.0 